% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineComboCounts.R
\name{combineComboCounts}
\alias{combineComboCounts}
\title{Combine combinatorial barcode counts}
\usage{
combineComboCounts(...)
}
\arguments{
\item{...}{Any number of \linkS4class{DataFrame}s produced by \code{\link{countComboBarcodes}}.}
}
\value{
A list containing:
\itemize{
\item \code{combination}, a DataFrame containing all unique combinatorial barcodes observed in any \code{...}.
Each row corresponds to a barcode and each column contains an integer identifier for the sequence in the variable region.
\item \code{counts}, a matrix with number of columns equal to number of objects in \code{...}. 
Each row corresponds to a unique combinatorial barcode in \code{keys} and each column represents the count of that barcode in each file of \code{files}.
}
}
\description{
Combine counts for combinatorial barcodes from multiple files into a single count matrix.
}
\examples{
df1 <- DataFrame(combination=I(DataFrame(X=1:4, Y=1:4)),
   count=sample(10, 4))

df2 <- DataFrame(combination=I(DataFrame(X=1:4, Y=4:1)),
   count=sample(10, 4))

df3 <- DataFrame(combination=I(DataFrame(X=1, Y=1)),
   count=sample(10, 1))

combineComboCounts(df1, df2, df3)

}
\author{
Aaron Lun
}
