% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countFixedBarcodes.R
\name{countFixedBarcodes}
\alias{countFixedBarcodes}
\title{Count single barcodes}
\usage{
countFixedBarcodes(fastq, choices, template = NULL,
  substitutions = TRUE, deletions = TRUE)
}
\arguments{
\item{fastq}{String containing the path to a FASTQ file containing single-end data,
or a connection object to such a file.}

\item{choices}{A \linkS4class{DataFrame} of sequences for the variable regions.
Each row should correspond to a barcode and each column should contain a character vector of sequences.}

\item{template}{A template for the barcode structure, see \code{\link{?createBarcodes}} for details.}

\item{substitutions}{Logical scalar specifying whether substitutions should be allowed when matching to variable regions.}

\item{deletions}{Logical scalar specifying whether deletions should be allowed when matching to variable regions.}
}
\value{
An integer vector of length equal to \code{nrow(choices)},
containing the frequency of each barcode.
This is named with the row names of \code{choices}.
}
\description{
Count the frequency of barcodes in a FASTQ file containing data for a single-end sequencing screen.
}
\details{
In the simplest case, \code{choices} can be specified with a single column of variable sequences.
If \code{template=NULL}, the function will then directly use each of those sequences as the barcode with no constant regions.
More complex barcode structures are accommodated via \code{template}, e.g., with multiple variable regions and intervening and/or flanking constant spacers.
See \code{\link{createBarcodes}} for more details.

If \code{substitutions=TRUE}, only one mismatch is allowed across all variable regions,
\emph{not} per variable region.
Similarly, if \code{deletions=TRUE}, only one deletion is allowed across all variable regions.
If both are set, only one deletion or mismatch is allowed across all variable regions,
i.e., there is a maximum edit distance of 1 from any possible reference combination.
}
\examples{
# Creating an example dual barcode sequencing experiment.
known.pool <- c("AGAGAGAGA", "CTCTCTCTC",
    "GTGTGTGTG", "CACACACAC")

N <- 1000
barcodes <- sprintf("CAGCTACGTACG\%sCCAGCTCGATCG",
   sample(known.pool, N, replace=TRUE))
names(barcodes) <- seq_len(N)

library(Biostrings)
tmp <- tempfile(fileext=".fastq")
writeXStringSet(DNAStringSet(barcodes), filepath=tmp, format="fastq")

# Counting the combinations.
countFixedBarcodes(tmp, choices=DataFrame(known.pool))

}
\author{
Aaron Lun
}
