% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countSingleBarcodes.R
\name{countSingleBarcodes}
\alias{countSingleBarcodes}
\title{Count single barcodes}
\usage{
countSingleBarcodes(fastq, choices, flank5, flank3, template = NULL,
  substitutions = TRUE, deletions = TRUE)
}
\arguments{
\item{fastq}{String containing the path to a FASTQ file containing single-end data,
or a connection object to such a file.}

\item{choices}{A character vector of sequences for the variable regions, one per barcode.}

\item{flank5}{String containing the constant sequence on the 5' flank of the variable region.}

\item{flank3}{String containing the constant sequence on the 3' flank of the variable region.}

\item{template}{A template for the barcode structure, see \code{\link{?createBarcodes}} for details.}

\item{substitutions}{Logical scalar specifying whether substitutions should be allowed when matching to variable regions.}

\item{deletions}{Logical scalar specifying whether deletions should be allowed when matching to variable regions.}
}
\value{
An integer vector of length equal to \code{nrow(choices)},
containing the frequency of each barcode.
This is named with the row names of \code{choices}.
}
\description{
Count the frequency of barcodes in a FASTQ file containing data for a single-end sequencing screen.
}
\details{
If \code{template} is specified, it will be used to define the flanking regions.
Any user-supplied values of \code{flank5} and \code{flank3} will be ignored.
Note that, for this function, the template should only contain a single variable region.
See \code{\link{parseBarcodeTemplate}} for more details.

If \code{substitutions=TRUE}, only one mismatch is allowed across all variable regions,
\emph{not} per variable region.
Similarly, if \code{deletions=TRUE}, only one deletion is allowed across all variable regions.
If both are set, only one deletion or mismatch is allowed across all variable regions,
i.e., there is a maximum edit distance of 1 from any possible reference combination.
}
\examples{
# Creating an example dual barcode sequencing experiment.
known.pool <- c("AGAGAGAGA", "CTCTCTCTC",
    "GTGTGTGTG", "CACACACAC")

N <- 1000
barcodes <- sprintf("CAGCTACGTACG\%sCCAGCTCGATCG",
   sample(known.pool, N, replace=TRUE))
names(barcodes) <- seq_len(N)

library(Biostrings)
tmp <- tempfile(fileext=".fastq")
writeXStringSet(DNAStringSet(barcodes), filepath=tmp, format="fastq")

# Counting the combinations.
countSingleBarcodes(tmp, choices=known.pool,
    template="CAGCTACGTACGNNNNNNNNNCCAGCTCGATCG")

countSingleBarcodes(tmp, choices=known.pool,
    flank5="CAGCTACGTACG", flank3="CCAGCTCGATCG")

}
\author{
Aaron Lun
}
