% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBarcodes.R
\name{createBarcodes}
\alias{createBarcodes}
\title{Create barcodes from a template}
\usage{
createBarcodes(template, choices)
}
\arguments{
\item{template}{String containing template sequence of a barcode.
Variable regions should be marked with N's.}

\item{choices}{A \linkS4class{DataFrame} of potential choices for the variable regions.
Each row represents a distinct barcode, and each column represents one of the variable regions
(in their order on \code{template}, from 5' to 3').}
}
\value{
Character vector containing realized barcodes.
This is named with the row names of \code{choices}.
}
\description{
Create barcodes from a template template by filling in N's with possible sequences.
}
\details{
The choices for the variable regions do not have to have the same length as in \code{template}.
This potentially allows for variable-length barcodes, if such a thing is desirable.
}
\examples{
# Single spacer:
createBarcodes("AAAANNNNCCCCNNGGGG", 
   DataFrame(c("T", "TT", "TTT")))

# Double spacer:
createBarcodes("AAAANNNNCCCCNNGGGG", 
   DataFrame(c("T", "TT", "TTT"), c("TTTT","TT", "T")))
}
\author{
Aaron Lun
}
