% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterReference.R
\name{.filterReference}
\alias{.filterReference}
\title{Filter on the reference}
\usage{
.filterReference(ref.field, to.use)
}
\arguments{
\item{ref.field}{String containing the name of the column of \code{colData} specifying the type of each sample.}

\item{to.use}{Character vector containing the types of samples that are references.}
}
\value{
A string containing commands to filter barcodes based on their abundances in the reference samples.
}
\description{
Generate code to compute a filter threshold for removing low-abundance barcodes,
based on the average log-abundance of all barcodes in the reference samples.
}
\details{
The output commands assume that there is a \linkS4class{SummarizedExperiment} object named \code{se} and a DGEList object named \code{y} in the evaluation environment.
Filtering will be applied based on the reference samples to modify \code{y} in place.
}
\examples{
cat(filterReference("condition", "ref"))
cat(filterReference("time", 0))

}
\seealso{
\code{\link{runVoomScreen}}, in which this function is called.
}
\author{
Aaron Lun
}
