% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsScreenAlignment.R
\name{ngsComboBarcodingAlignment}
\alias{ngsComboBarcodingAlignment}
\title{Process reads for cell barcoding experiment}
\usage{
ngsComboBarcodingAlignment(
  ngs,
  choices,
  template,
  runs = runs,
  aggregateBy = c("samid", "libid"),
  ...
)
}
\arguments{
\item{ngs}{String specyfing the NGS number (eg. "ngs3096").}

\item{choices}{A \linkS4class{List} of character vectors, one per variable region in \code{template}.
The first vector should contain the potential sequences for the first variable region, 
the second vector for the second variable region and so on.}

\item{template}{A template for the barcode structure, see \code{?\link{parseBarcodeTemplate}} for details.}

\item{runs}{Character string indicating which sequencing runs should be considered.}

\item{aggregateBy}{String specifying which id (samid or libid) should be used as sample unit.}

\item{...}{Further arguments to pass to \code{matrixOfSingleBarcodes}.}
}
\value{
\itemize{
\item An integer matrix named \code{"counts"}, where each column corresponds to a SAMID.
\item Row metadata containing the barcodes information
\item Column metadata containing sample information as annotated in HiTS-LIMS, as well as;
an integer vector \code{nreads}, containing the total number of reads in each file;
and \code{nmapped}, containing the number of reads assigned to a combo barcode in; 
the output count matrix.
}
}
\description{
Process reads for cell barcoding experiment using lentiviral delivery.
}
\examples{
# Processing a tumor barcoding experiment using Cellecta50M lib:
\dontrun{
library(crisprAnnotation)
barcodes <- getBarcodeLibrary("barcodes.cellecta.celltracker.50M")
varRegion <- paste0(rep("N",18), collapse="")
template <- paste0("GTTCG", 
   varRegion,"TTCG",varRegion, 
   "TTCGG"
)
se <- getBarcodeCounts(ngs, list(barcodes, barcodes), template)
}

}
\author{
Jean-Philippe Fortin
}
