% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsScreenAlignment.R
\name{ngsScreenAlignment}
\alias{ngsScreenAlignment}
\title{Process reads from CRISPR and ORF screens}
\usage{
ngsScreenAlignment(ngs, libname, flank5, flank3, runs = NULL, ...)
}
\arguments{
\item{ngs}{String specyfing the NGS number (eg. "ngs3096").}

\item{libname}{String specifying the name of the CRISPR/ORF library. Must be available in crisprAnnotation.}

\item{flank5}{String containing the constant sequence on the 5' flank of the variable region.}

\item{flank3}{String containing the constant sequence on the 3' flank of the variable region.}

\item{runs}{Character string indicating which sequencing runs should be considered.}

\item{...}{Further arguments to pass to \code{matrixOfSingleBarcodes}.}
}
\value{
\itemize{
\item An integer matrix named \code{"counts"}, where each column corresponds to a SAMID.
\item Row metadata containing the CRISPR library annotation.
\item Column metadata containing sample information as annotated in HiTS-LIMS, as well as;
an integer vector \code{nreads}, containing the total number of reads in each file;
and \code{nmapped}, containing the number of reads assigned to a barcode in the output count matrix.
}
}
\description{
Process reads from CRISPR and ORF screens and return a SummarizedExperiment object.
}
\examples{
# Processing CRISPRa screen NGS3096 (Pilot screen with Transylvania library)
\dontrun{
ngsScreenAlignment(ngs="ngs3096",
    libname="crispra.cas9.human.transylvania1",
    flank5="TACCG", flank3="GTTT"
)
}
}
\author{
Jean-Philippe Fortin
}
