% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeControls.R
\name{.normalizeControls}
\alias{.normalizeControls}
\title{Normalization on controls}
\usage{
.normalizeControls(type.field, to.use)
}
\arguments{
\item{type.field}{String containing the name of the column of \code{rowData} specifying the type of the gene targeted by each barcode.}

\item{to.use}{Character vector containing the types to use during normalization.}
}
\value{
A string containing commands to compute normalization factors based on the specified subset of barcodes.
}
\description{
Generate code to perform median-based normalization on a subset of features.
usually barcodes for control genes to satisfy the assumption of a majority of non-DE features.
}
\details{
The output commands assume that there is a \linkS4class{SummarizedExperiment} object named \code{se} and a DGEList object named \code{y} in the evaluation environment.
Normalization factors are computed using TMM normalization on the subset of control barcodes,
and assigned back to \code{y} for downstream analysis.
}
\examples{
cat(normalizeControls("gene.type", "NTG"))
cat(normalizeControls("gene.type", "NEG"))

# One can also specify multiple features,
# but this is probably unwise.
cat(normalizeControls("gene.type", c("NTG", "NEG")))

}
\author{
Aaron Lun
}
