% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVoomScreen.R
\name{runVoomScreen}
\alias{runVoomScreen}
\title{Use \code{voom} to analyze a screen}
\usage{
runVoomScreen(se, groups, comparisons, covariates = NULL, block = NULL,
  ..., type.field = "XXX", norm.types = NULL, design.fun = NULL,
  contrasts.fun = NULL, fname = NULL)
}
\arguments{
\item{se}{A \linkS4class{SummarizedExperiment} object.
Alternatively, a string containing a file path to a serialized instance of a SummarizedExperiment.}

\item{groups}{String specifying the column of \code{colData(se)} containing the grouping factor, see \code{\link{createDesignMatrix}}.}

\item{comparisons}{A list of character vectors specifying the comparisons to perform between groups, see \code{\link{createContrasts}}.}

\item{covariates}{Character vector specifying the columns of \code{colData(se)} containing continuous covariates of interest,  see \code{\link{createDesignMatrix}}.}

\item{block}{Character vector specifying additional blocking factors or covariates that are \emph{not} of interest, see \code{\link{createDesignMatrix}}.}

\item{...}{Further arguments to pass to \code{\link{runVoom}}.}

\item{type.field}{String specifying the field of \code{rowData} specifying the gene type for each barcode.}

\item{norm.types}{Character vector specifying the gene types on which to perform normalization.}

\item{design.fun}{A function to create a custom design matrix, see \code{?\link{createDesignMatrix}}.}

\item{contrasts.fun}{A list of custom contrasts information, including contrast-generating functions; see \code{?\link{createContrasts}}.}

\item{fname}{String containing the path to an output Rmarkdown file.}
}
\value{
A list containing \code{objects} and \code{results}.
\code{objects} is a list containing the \code{fit} object, the MArrayLM object produced by \code{eBayes}.
\code{results} is a named list containing \linkS4class{DataFrame}s of result tables from all contrasts.
}
\description{
Perform a \code{voom} analysis on a count matrix from a sequencing screen to detect differentially abundant barcodes across samples.
}
\details{
This function is largely a wrapper around \code{\link{runVoom}}, with two additional features:
\itemize{
\item Normalization based on non-targeting genes (NTGs) and/or non-essential genes (NEGs).
These are negative controls that may yield more accurate normalization factors when the majority of other genes are DE.
\item Consolidation of per-barcode results into per-gene results, in experiments where multiple barcodes (i.e., guides, shRNAs) target the same gene.
}
}
\examples{
# Mocking up an example dataset.
N <- 1000
mu <- 2^runif(N, 2, 10)
y <- matrix(rnbinom(N * 6, mu=mu, size=10), ncol=6)
g <- gl(2, 3)

library(SummarizedExperiment)
se <- SummarizedExperiment(y)
rowData(se)$type <- sample(c("endog", "NTG", "NEG"), N, replace=TRUE)
colData(se)$group <- g

out <- runVoomScreen(se, groups="group",
    comparisons=list(c("2", "1")),
    type.field="type", norm.types="NEG")
head(out$results$`2-1`)

}
\author{
Aaron Lun
}
