% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreenBarcodeStatFrame.R
\docType{class}
\name{ScreenBarcodeStatFrame}
\alias{ScreenBarcodeStatFrame}
\alias{ScreenBarcodeStatFrame-class}
\alias{.trackCheck,ScreenBarcodeStatFrame-method}
\alias{.trackRowIds,ScreenBarcodeStatFrame-method}
\title{DataFrame for barcode-level screen results}
\description{
The ScreenBarcodeStatFrame class is literally identical to a standard \linkS4class{DataFrame},
and can be used as such in all applications.
It is intended to hold barcode-level results from a differential abundance analysis of barcode sequencing data from high-throughput CRISPR/siRNA screens,
where each row corresponds to a separate barcode.
}
\section{Constructor}{

\code{ScreenBarcodeStatFrame(x, ...)} will return a ScreenBarcodeStatFrame object, given the arguments:
\itemize{
\item \code{x}, a \linkS4class{DataFrame} object or something that can be coerced into one.
\item \code{...}, other named fields to add to the provenance tracking information via \code{\link{trackinfo}}.
}
}

\section{Checking metadata}{

\code{\link{.trackCheck}(x)} will check for the presence of correct provenance fields in a ScreenBarcodeStatFrame \code{x}
and return a list of this information.
Required fields include:
\itemize{
\item \code{origin}, the sources from which \code{x} was derived - see \linkS4class{TrackedDataFrame}.
\item \code{type}, the type of result - see \linkS4class{TrackedDataFrame}.
\item \code{description}, a plain-English description - see \linkS4class{TrackedDataFrame}.
\item \code{design}, the numeric design matrix - see \linkS4class{DiffStatFrame}.
\item \code{contrast}, the contrast vector/matrix - see \linkS4class{DiffStatFrame}.
\item \code{method}, a string specifying the DE analysis method that was used - see \linkS4class{DGEStatFrame}.
}
A \code{type} field specifying the \dQuote{per-barcode differential screen abundance} result type in a controlled vocabulary is included in the output,
and will overwrite any \code{type} field in \code{trackinfo(x)}.
}

\section{Checking column names}{

\code{\link{.trackCheck}(x)} will check that the column names of \code{x} include:
\itemize{
\item \code{"PValue"}, a numeric field containing the p-value for each barcode.
\item \code{"FDR"}, a numeric field containing the Benjamini-Hochberg adjusted p-value for each barcode.
\item \code{"AveAb"}, a numeric field containing the average abundance of each barcode in log2-counts-per-million.
}
It should also contain either one \code{"LogFC"} field or multiple \code{"LogFC."}-prefixed fields.
These contain the log2-fold changes for a single contrast vector or an ANOVA-like contrast matrix, respectively.

The feature corresponding to each barcode is not mandatory here, as this may not exist for some experiments.
If it does exist, the mapping can be determined from the data and need not be stored in the result Frame.
Nonetheless, pipelines may consider adding the feature name for user convenience.
}

\section{Checking row names}{

\code{\link{.trackCheck}(x)} will also check that the row names of \code{x} are a subset of the permissible feature names.
These globally permissible names are automatically added upon importing a \linkS4class{SummarizedExperiment} with \pkg{gp.sa.io} functions.
(In rare cases, users may define the permissible names themselves with \code{\link{.setDimnames}} but this should not be a routine practice.)
}

\examples{
library(gp.sa.core)

# Mocking up the aftermath of a DE analysis:
de.output <- DataFrame(LogFC=1:10, AveAb=1:10, 
   PValue=0:9/10, FDR=0:9/10)

Y <- ScreenBarcodeStatFrame(de.output, 
    design=cbind(A=c(X=1, Y=-1), B=2),
    contrast=c(A=1, B=-1), feature="barcode",
    method="voom", description="I did voom")
Y 

}
\seealso{
\linkS4class{ScreenStatFrame}, from which this class is derived.
}
\author{
Aaron Lun
}
