% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreenFeatureStatFrame.R
\docType{class}
\name{ScreenFeatureStatFrame}
\alias{ScreenFeatureStatFrame}
\alias{ScreenFeatureStatFrame-class}
\alias{.trackCheck,ScreenFeatureStatFrame-method}
\alias{.trackReqCols,ScreenFeatureStatFrame-method}
\alias{.trackRowIds,ScreenFeatureStatFrame-method}
\title{DataFrame for feature-level screen results}
\description{
The ScreenFeatureStatFrame class is literally identical to a standard \linkS4class{DataFrame},
and can be used as such in all applications.
It is intended to hold feature-level results from a differential abundance analysis of barcode sequencing data from high-throughput CRISPR/siRNA screens,
where each row corresponds to a separate feature (usually a gene for which there are multiple barcodes).
}
\section{Constructor}{

\code{ScreenFeatureStatFrame(x, ...)} will return a ScreenFeatureStatFrame object, given the arguments:
\itemize{
\item \code{x}, a \linkS4class{DataFrame} object or something that can be coerced into one.
\item \code{...}, other named fields to add to the provenance tracking information via \code{\link{trackinfo}}.
}
}

\section{Checking metadata}{

\code{\link{.trackCheck}(x)} will check for the presence of correct provenance fields in a ScreenFeatureStatFrame \code{x}
and return a list of this information.
Required fields include:
\itemize{
\item \code{origin}, the sources from which \code{x} was derived - see \linkS4class{TrackedDataFrame}.
\item \code{type}, the type of result - see \linkS4class{TrackedDataFrame}.
\item \code{description}, a plain-English description - see \linkS4class{TrackedDataFrame}.
\item \code{design}, the numeric design matrix - see \linkS4class{DiffStatFrame}.
\item \code{contrast}, the contrast vector/matrix - see \linkS4class{DiffStatFrame}.
\item \code{method}, a string specifying the DE analysis method that was used - see \linkS4class{DGEStatFrame}.
}
A \code{type} field specifying the \dQuote{feature-level differential screen abundance} result type in a controlled vocabulary is included in the output,
and will overwrite any \code{type} field in \code{trackinfo(x)}.
}

\section{Checking column names}{

\code{\link{.trackCheck}(x)} will check that the column names of \code{x} include:
\itemize{
\item \code{"NBarcodes"}, an integer field containing the number of barcodes for each feature.
\item \code{"PValue"}, a numeric field containing the p-value for each feature.
\item \code{"FDR"}, a numeric field containing the Benjamini-Hochberg adjusted p-value for each feature.
\item \code{"AveAb"}, a numeric field containing the average abundance of each feature in log2-counts-per-million.
}
It should also contain either one \code{"LogFC"} field or multiple \code{"LogFC."}-prefixed fields.
These contain the log2-fold changes for a single contrast vector or an ANOVA-like contrast matrix, respectively.

In practice, all of these statistics are usually obtained by combining the statistics for all barcodes assigned to a given feature.
For example, the log-fold changes and average abundances for each feature are usually values from some appropriately representative barcode.
Refer to \code{\link{barcodeSetTest}} and \code{\link{combineBarcodeTests}} for more details.
}

\examples{
library(gp.sa.core)

# Mocking up the aftermath of a DE analysis:
de.output <- DataFrame(NBarcodes=10, LogFC=1:10, AveAb=1:10, 
   PValue=0:9/10, FDR=0:9/10)

Y <- ScreenFeatureStatFrame(de.output, 
    design=cbind(A=c(X=1, Y=-1), B=2),
    contrast=c(A=1, B=-1), feature="barcode",
    method="voom", description="I did voom")
Y 

}
\seealso{
\linkS4class{ScreenStatFrame}, from which this class is derived.
}
\author{
Aaron Lun
}
