% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenNormalize.R
\name{tmmOnSubset}
\alias{tmmOnSubset}
\title{Subsetted TMM normalization}
\usage{
tmmOnSubset(type.field, to.use)
}
\arguments{
\item{type.field}{String containing the name of the column of \code{rowData} specifying the type of the gene targeted by each barcode.}

\item{to.use}{Character vector containing the types to use during normalization.
If \code{NULL}, all barcodes are used.}
}
\value{
A string containing commands to compute normalization factors based on the specified subset of barcodes.
}
\description{
Generate code to perform TMM normalization on a subset of features,
usually barcodes for control genes to satisfy the assumption of a majority of non-DE features.
}
\details{
The output commands assume that there is a \linkS4class{SummarizedExperiment} object named \code{se} and a DGEList object named \code{y} in the evaluation environment.
The subsetted normalization factors will be assigned back to \code{y} for downstream analysis.
}
\examples{
cat(tmmOnSubset("gene.type", c("NTG", "NEG")))

}
\author{
Aaron Lun
}
