% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidateGenes.R
\name{.consolidateGenes}
\alias{.consolidateGenes}
\title{Consolidate to genes}
\usage{
.consolidateGenes(gene.field)
}
\arguments{
\item{gene.field}{String specifying the field of \code{rowData(se)} that contains the gene identifier for each barcode.}
}
\value{
A function that takes a machine-readable name of the contrast (see \code{\link{.createContrasts}}),
and returns a Rmarkdown chunk containing code to store per-barcode and per-gene results in the \code{all.results} list.
}
\description{
Consolidate barcode-level statistics to gene-level statistics.
}
\details{
The returned function needs a string equivalent to the \code{name} field in the DataFrame from \code{\link{.createContrasts}}.
This is used to determine the name of the output files to which to save results.

The Rmarkdown chunk expects the SummarizedExperiment \code{se}, a \code{DGEList} object \code{y} 
and a per-barcode result table \code{res} to be present in the evaluation environment.

Note that all results for a single contrast are saved as a single entry of \code{all.results};
per-barcode and per-gene results are accommodated by making this entry a list of length two to hold the separate result tables.
It is useful to unroll this list for easier parsing by users later, see \code{\link{runVoomScreen}}.
}
\examples{
FUN <- .consolidateGenes("gene.type")
cat(FUN("A-B"))
}
