% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidateGenes.R
\name{.consolidateGenes}
\alias{.consolidateGenes}
\title{Consolidate to genes}
\usage{
.consolidateGenes(gene.field)
}
\arguments{
\item{gene.field}{String specifying the field of \code{rowData(se)} that contains the gene identifier for each barcode.}
}
\value{
A function that takes a machine-readable name of the contrast (see \code{\link{.createContrasts}}),
and returns a Rmarkdown chunk containing code to:
\itemize{
\item store per-barcode results in the \code{all.results} list, named with a \code{:barcode} suffix.
\item combine barcode-level statistics for that contrast into gene-level statistics.
\item store per-gene results in the \code{all.results} list, named with a \code{:gene} suffix.
}
}
\description{
Consolidate barcode-level statistics to gene-level statistics.
}
\details{
The returned function needs a string equivalent to the \code{name} field in the DataFrame from \code{\link{.createContrasts}}.
This is used to determine the name of the output files to which to save results.

The Rmarkdown chunk expects the SummarizedExperiment \code{se}, a \code{DGEList} object \code{y} and a result table \code{res} to be present in the evaluation environment.
}
\examples{
FUN <- .consolidateGenes("gene.type")
cat(FUN("A-B"))
}
