% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidateGenes.R
\name{consolidateGenes}
\alias{consolidateGenes}
\title{Consolidate to genes}
\usage{
consolidateGenes(res.dir, gene.field)
}
\arguments{
\item{gene.field}{String specifying the field of \code{rowData(se)} that contains the gene identifier for each barcode.}
}
\value{
A function that takes a machine-readable name of the contrast (see \code{\link{createContrasts}}),
and returns a Rmarkdown chunk containing code to:
\itemize{
\item save per-barcode results to file, with a \code{:barcode.csv} or \code{:barcode.rds} suffix.
\item combine barcode-level statistics for that contrast into gene-level statistics.
\item save per-gene results to file, with a \code{:gene.csv} or \code{:gene.rds} suffix.
}
}
\description{
Consolidate barcode-level statistics to gene-level statistics.
}
\details{
We use Simes' method as it is fast, statistically rigorous and able to detect genes with only a minority of differentially abundant barcodes.
The exact implementation uses the \code{\link{combineTests}} function from the \pkg{csaw} package.
We also report the abundance and log-fold changes of the best barcode within each gene.

The returned function needs a string equivalent to the \code{name} field in the DataFrame from \code{\link{createContrasts}}.
This is used to determine the name of the output files to which to save results.

The Rmarkdown chunk expects the SummarizedExperiment \code{se}, a \code{DGEList} object \code{y} and a result table \code{res} to be present in the evaluation environment.
}
\examples{
FUN <- consolidateGenes("gene.type")
cat(FUN(DataFrame(name="A-B")))
}
