% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countSingleBarcodes.R
\name{countSingleBarcodes}
\alias{countSingleBarcodes}
\title{Count single barcodes}
\usage{
countSingleBarcodes(fastq, choices, template = "N")
}
\arguments{
\item{fastq}{String containing the path to a FASTQ file containing single-end data,
or a connection object to such a file.}

\item{choices}{A \linkS4class{DataFrame} of potential barcodes.
Each row should correspond to a barcode and each column should contain a character vector of sequences.}

\item{template}{A template for the barcode structure, see \code{\link{?createBarcodes}} for details.}
}
\value{
An integer vector of length equal to \code{nrow(choices)},
containing the frequency of each barcode.
This is named with the row names of \code{choices}.
}
\description{
Count the frequency of barcodes in a FASTQ file containing data for a single-end sequencing screen.
}
\details{
In the simplest case, only \code{choices} needs to be specified with the barcode sequences.
However, this function can accommodate more complex barcode structures via \code{template},
e.g., with multiple variable regions and intervening spacers.
See \code{\link{createBarcodes}} for more details.
}
\examples{
# Re-using ShortRead's examples:
library(ShortRead)
sp <- SolexaPath(system.file('extdata', package='ShortRead'))
fl <- file.path(analysisPath(sp), "s_1_sequence.txt")

countSingleBarcodes(fl,
   choices=DataFrame(c("AAA", "GGG", "CCC", "TTT")))

}
\author{
Aaron Lun
}
