% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVoomScreen.R
\name{runVoomScreen}
\alias{runVoomScreen}
\title{Use \code{voom} to analyze a screen}
\usage{
runVoomScreen(se, groups, comparisons, covariates = NULL, block = NULL,
  ..., reference.field, reference.level, norm.type.field, norm.type.level,
  gene.field, design.fun = NULL, contrasts.fun = NULL, fname = NULL)
}
\arguments{
\item{se}{A \linkS4class{SummarizedExperiment} object.
Alternatively, a string containing a file path to a serialized instance of a SummarizedExperiment.}

\item{groups}{String specifying the column of \code{colData(se)} containing the grouping factor, see \code{\link{createDesignMatrix}}.}

\item{comparisons}{A list of character vectors specifying the comparisons to perform between groups, see \code{\link{createContrasts}}.}

\item{covariates}{Character vector specifying the columns of \code{colData(se)} containing continuous covariates of interest,  see \code{\link{createDesignMatrix}}.}

\item{block}{Character vector specifying additional blocking factors or covariates that are \emph{not} of interest, see \code{\link{createDesignMatrix}}.}

\item{...}{Further arguments to pass to \code{\link{runVoom}}.}

\item{reference.field}{String specifying the column of \code{colData(se)} containing the type of each sample (i.e., reference or not).}

\item{reference.level}{Character vector specifying the reference levels of the column named by \code{reference.field}.}

\item{norm.type.field}{String specifying the field of \code{rowData(se)} containing the gene type for each barcode.}

\item{norm.type.level}{Character vector specifying the gene types on which to perform normalization.}

\item{gene.field}{String specifying the field of \code{rowData(se)} that contains the gene identifier for each barcode.}

\item{design.fun}{A function to create a custom design matrix, see \code{?\link{createDesignMatrix}}.}

\item{contrasts.fun}{A list of custom contrasts information, including contrast-generating functions; see \code{?\link{createContrasts}}.}

\item{fname}{String containing the path to an output Rmarkdown file.}
}
\value{
A list containing \code{objects} and \code{results}.
\code{objects} is a list containing the \code{fit} object, the MArrayLM object produced by \code{eBayes}.
\code{results} is a named list containing \linkS4class{DataFrame}s of result tables from all contrasts.
}
\description{
Perform a \code{voom} analysis on a count matrix from a sequencing screen to detect differentially abundant barcodes across samples.
}
\details{
This function is largely a wrapper around \code{\link{runVoom}}, with three additional features:
\itemize{
\item Filtering based on reference samples; see \code{\link{filterReference}}.
If \code{reference.field=NULL}, default edgeR filtering is used instead; see \code{\link{defaultEdgeRFilter}}.
If \code{reference.field=NA}, no filtering is performed.
\item Normalization based on non-targeting genes (NTGs) and/or non-essential genes (NEGs); see \code{\link{normalizeControls}}.
If \code{norm.type.field=NULL}, default edgeR normalization is used instead; see \code{\link{defaultEdgeRNormalize}}.
If \code{norm.type.field=NA}, no normalization is performed beyond library size normalization.
\item Consolidation of per-barcode results into per-gene results, in experiments where multiple barcodes (i.e., guides, shRNAs) target the same gene; see \code{\link{consoliateGenes}}.
If \code{gene.field=NA}, no consolidation is performed.
}
}
\examples{
# Mocking up an example dataset.
N <- 1000
mu <- 2^runif(N, 2, 10)
y <- matrix(rnbinom(N * 6, mu=mu, size=10), ncol=6)
g <- gl(2, 3)

library(SummarizedExperiment)
se <- SummarizedExperiment(y)
rowData(se)$type <- sample(c("endog", "NTG", "NEG"), N, replace=TRUE)
rowData(se)$gene <- paste0("GENE_", sample(100, N, replace=TRUE))
colData(se)$group <- g

tmp <- tempfile(fileext=".Rmd")
out <- runVoomScreen(se, groups="group",
    comparisons=list(c("2", "1")),
    norm.type.field="type", norm.type.level="NEG",
    reference.field="group", reference.level="1",
    gene.field="gene", fname=tmp)
file.exists(tmp)

head(out$results$`2-1_barcode`)
head(out$results$`2-1_gene`)

}
\author{
Aaron Lun
}
