% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVoomScreen.R
\name{runVoomScreen}
\alias{runVoomScreen}
\title{Use \code{voom} to analyze a screen}
\usage{
runVoomScreen(
  se,
  groups,
  comparisons,
  reference.field,
  reference.level,
  norm.type.field,
  norm.type.level,
  gene.field,
  method = c("simes", "holm-mid", "fry"),
  ...,
  annotation = NULL,
  lfc = 0,
  robust = TRUE,
  dup.cor = NULL,
  contrasts.fun = NULL,
  dump.norm = NULL,
  fname = "voom-screen.Rmd",
  commit = "auto",
  save.all = TRUE
)
}
\arguments{
\item{se}{A \linkS4class{SummarizedExperiment} object containing read counts for each barcode (row) and sample (column).
Alternatively, a database ID string or an \linkS4class{InputResource} object, see \code{?"\link{gp.sa.core-data-inputs}"}.}

\item{groups}{String specifying the column of \code{colData(se)} containing the grouping factor, see \code{?"\link[gp.sa.diff]{gp.sa.diff-args}"}.}

\item{comparisons}{A list of character vectors specifying simple contrasts to perform between groups, see \code{?"\link[gp.sa.diff]{gp.sa.diff-args}"}.}

\item{reference.field}{String specifying the column of \code{colData(se)} containing the type of each sample (i.e., reference or not).}

\item{reference.level}{Character vector specifying the reference levels of the column named by \code{reference.field}.}

\item{norm.type.field}{String specifying the field of \code{rowData(se)} containing the gene type for each barcode.}

\item{norm.type.level}{Character vector specifying the gene types on which to perform normalization.}

\item{gene.field}{String specifying the field of \code{rowData(se)} that contains the gene identifier for each barcode.}

\item{method}{String specifying the consolidation method to convert per-barcode statistics into per-gene results.}

\item{...}{Further arguments to pass to \code{\link{runVoom}}.}

\item{annotation}{Character vector specifying feature annotation to carry through into the result table.}

\item{lfc}{Numeric scalar specifying a log-fold change threshold to be used in \code{\link[gp.sa.diff]{treat}}.
A value of zero will default to the standard analysis with \code{\link[gp.sa.diff]{eBayes}}.}

\item{robust}{Logical scalar indicating whether robust empirical Bayes shrinkage should be used in \code{\link[gp.sa.diff]{eBayes}}.}

\item{dup.cor}{String specifying the blocking factor to use in \code{\link[gp.sa.diff]{duplicateCorrelation}}.
No additional blocking is performed if \code{NULL}.}

\item{contrasts.fun}{A list of custom contrasts information, see \code{?"\link[gp.sa.diff]{gp.sa.diff-args}"}.}

\item{dump.norm}{String specifying a path to an output file to save normalized abundances in a CSV file.
This file is intended only for diagnostic inspection and should \emph{not} be used as input into further GPSA pipeline.}

\item{fname}{String containing the path to an output Rmarkdown file.}

\item{commit}{String specifying the auto-committing behavior, see \code{?"\link[gp.sa.diff]{gp.sa.core-auto-commits}"}.}

\item{save.all}{Logical scalar indicating whether the returned \linkS4class{DAScreenStatFrame}s should also be saved to file.
Ignored if \code{se} lacks provenance information, in which case saving is never performed.}
}
\value{
A \linkS4class{List} containing two Lists, \code{barcode} and \code{gene}.
Each list contains barcode- and gene-level result tables as \linkS4class{DAScreenStatFrame}s from all contrasts.
If \code{gene.field=NA}, only the \code{barcode} List is returned.

A Rmarkdown file is also created at \code{fname}, containing the steps required to reproduce the analysis.
This also provides a basis for further customization.
}
\description{
Perform a \code{voom} analysis on a count matrix from a sequencing screen to detect differentially abundant barcodes across samples.
}
\details{
This function is largely a wrapper around \code{\link{runVoom}}, with three additional features:
\itemize{
\item Filtering based on reference samples, see below.
If \code{reference.field=NULL}, default edgeR filtering is used instead based \code{\link{filterByExpr}}.
If \code{reference.field=NA}, no filtering is performed.
\item Normalization based on non-targeting genes (NTGs) and/or non-essential genes (NEGs), see below.
If \code{norm.type.field=NULL}, default edgeR normalization is used instead based on \code{\link{calcNormFactors}}.
If \code{norm.type.field=NA}, no normalization is performed beyond library size normalization.
\item Consolidation of per-barcode results into per-gene results, in experiments where multiple barcodes (i.e., guides, shRNAs) target the same gene; see below.
If \code{gene.field=NA}, no consolidation is performed.
}
}
\section{Filtering on reference samples}{

We compute the log-average CPM across reference samples for each barcode.
We define a filtering threshold based on the median absolute deviation (MAD) below the median of the log-average CPMs across barcodes.
Barcodes that have log-average CPMs that are more than three MADs below the median are discarded,
as these correspond to barcodes that were most likely missing from the original pool during its manufacture.

Technically, this filtering strategy is not independent as it can theoretically enrich for barcodes that are upregulated in the reference.
In practice, this is not an issue as the filtering is only designed to remove outliers.
The density of barcodes around the filter boundary is so low that there will be no noticeable effect on type I error.
}

\section{Normalization with control barcodes}{

Restricting the normalization to only use control barcodes weakens the assumption of a non-DE majority of barcodes.
We currently use TMM normalization (i.e., robust average of a ratio) rather than methods based on a robust average of expression within each sample.
This avoids inflated errors due to the greater variance of the distribution of expression values compared to the ratio.
}

\section{Consolidating barcodes to genes}{

When \code{method="holm-mid"}, we use the \dQuote{minimum-Holm} method in \code{\link{combineBarcodeTests}} to combine p-values from multiple barcodes into a single p-value per gene.
This applies a Holm-Bonferroni correction across all barcodes for each gene before taking the \eqn{k}-th largest p-value for that gene.
The idea is to detect genes for which at least \eqn{k} barcodes are significantly differentially abundant, requiring some level of agreement between barcodes to reduce the influence of off-target guides.
We also report the log-counts-per-million and log-fold changes of the best barcode within each gene.

When \code{method="simes"}, we use Simes' method in \code{\link{combineBarcodeTests}} to combine barcode-level p-values.
This is fast, statistically rigorous and robust to correlations between guides for the same gene.
It is able to detect genes with only a minority of differentially abundant barcodes, though it will favor genes with many significant barcodes.
This tends to be the most sensitive approach but is also more susceptible to off-target effects where one outlier guide drives the gene-level result.
Again, we report the log-CPM and log-fold changes of the best barcode within each gene.

When \code{method="fry"}, we use the \code{\link{fry}} method to combine per-barcode statistics into a per-gene p-value.
This effectively applies gene set testing machinery on the set of guides for each gene.
The aim is to identify genes for which the mean t-statistic is significantly non-zero, which favors consistency among guides more than \code{method="simes"},
This may yield low p-values for genes with weak but consistent DE in each guide, at the cost of genes that have strong DE but only in a few guides.
Note that this mode is not compatible with ANOVA-like contrasts or with non-zero \code{lfc}.
}

\examples{
# Setting up a new project.
library(gp.sa.core)
proj <- tempfile()
newProject(proj, title="Aaron's project",
    description="This is one of Aaron's projects")

# Mocking up an example dataset.
N <- 1000
mu <- 2^runif(N, 2, 10)
y <- matrix(rnbinom(N * 6, mu=mu, size=10), ncol=6)
g <- gl(2, 3)

library(SummarizedExperiment)
se <- SummarizedExperiment(y)
rowData(se)$type <- sample(c("endog", "NTG", "NEG"), N, replace=TRUE)
rowData(se)$gene <- paste0("GENE_", sample(100, N, replace=TRUE))
colData(se)$group <- g

report <- file.path(proj, "voom-screen.Rmd")
out <- runVoomScreen(se, groups="group",
    comparisons=list(c("2", "1")),
    norm.type.field="type", norm.type.level="NEG",
    reference.field="group", reference.level="1",
    gene.field="gene", fname=report)

# Returns the Rmarkdown report and analysis results.
file.exists(report)
out
out[[1]]

}
\author{
Aaron Lun
}
